Bug fix: if there is no upload-tmp dir, error happens http://wordpress.org/support/topic/263861?replies=6#post-1053033
git-svn-id: https://plugins.svn.wordpress.org/contact-form-7/trunk@112567 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wp-contact-form-7.php
@@ -1553,17 +1553,21 @@ var _wpcf7 = {
     }
 
 	function cleanup_upload_files() {
-    $dir = $this->upload_tmp_dir();
-    $dir = trailingslashit($dir);
-    if ($handle = opendir($dir)) {
-      while (false !== ($file = readdir($handle))) {
-        if ($file == "." || $file == "..")
-          continue;
+        $dir = $this->upload_tmp_dir();
+        $dir = trailingslashit($dir);
+
+        if (! is_dir($dir))
+            return false;
+
+        if ($handle = opendir($dir)) {
+            while (false !== ($file = readdir($handle))) {
+                if ($file == "." || $file == "..")
+                    continue;
 
 				$stat = stat($dir . $file);
 				if ($stat['mtime'] + 60 < time()) // 60 secs
 					@ unlink($dir . $file);
-			}
+            }
 			closedir($handle);
 		}
 	}
